Added Task 2 - Security Needs of the Users
I didn't want to go overboard with this section and accidentally sign us up to test security features that aren't in the scope. That being said, I did include Network Security because I believe it is relevant due to the use of a web client to interact with the software. I would like to have the group look this over before we submit.