Updated Task 5, Contact with OSS team
This can be changed if we feel like it is inadequate. I wanted to have something in there so that everyone knows where we are at. If anyone HAS had contact with OpenMRS please send me the emails/transcripts or a description of the interaction and I can get that typed up and entered here.